parkstay: refactor campsite components
