+++ parkstay/frontend/parkstay/src/components/campsites/closureHistory/closeCampsite.vue
@@ -1,27 +1,17 @@
-<template id="addMaxStayCS">
-<bootstrapModal title="Add New Maximum Stay Period" :large=true @ok="addMaxStay()" okText="Add">
+<template id="pkCsClose">
+<bootstrapModal title="(Temporarily) close campsite" :large=true @ok="addClosure()">
 
     <div class="modal-body">
-        <form id="addMaxStayForm" class="form-horizontal">
+        <form id="closeCGForm" class="form-horizontal">
             <div class="row">
 			    <alert :show.sync="showError" type="danger"></alert>
                 <div class="form-group">
                     <div class="col-md-2">
-                        <label for="stay_maximum">Maximum Stay: </label>
+                        <label for="open_cg_range_start">Closure start: </label>
                     </div>
                     <div class="col-md-4">
-                        <input placeholder="Default = 28" id='stay_maximum' v-model="formdata.max_days" type='text' class="form-control" />
-                    </div>
-                </div>
-            </div>
-            <div class="row">
-                <div class="form-group">
-                    <div class="col-md-2">
-                        <label for="stay_start_picker">Period Start: </label>
-                    </div>
-                    <div class="col-md-4">
-                        <div class='input-group date' id="stay_start_picker">
-                            <input name="stay_start" v-model="formdata.range_start" type='text' class="form-control" />
+                        <div class='input-group date' id='close_cg_range_start'>
+                            <input  name="closure_start" v-model="formdata.range_start" type='text' class="form-control" />
                             <span class="input-group-addon">
                                 <span class="glyphicon glyphicon-calendar"></span>
                             </span>
@@ -32,11 +22,11 @@
             <div class="row">
                 <div class="form-group">
                     <div class="col-md-2">
-                        <label for="stay_end_picker">Period End: </label>
+                        <label for="open_cg_range_start">Closure end: </label>
                     </div>
                     <div class="col-md-4">
-                        <div class='input-group date' id='stay_end_picker'>
-                            <input name="stay_end" v-model="formdata.range_end" type='text' class="form-control" />
+                        <div class='input-group date' id='close_cg_range_end'>
+                            <input name="closure_end" v-model="formdata.range_end" type='text' class="form-control" />
                             <span class="input-group-addon">
                                 <span class="glyphicon glyphicon-calendar"></span>
                             </span>
@@ -47,14 +37,13 @@
             <div class="row">
                 <div class="form-group">
                     <div class="col-md-2">
-                        <label for="stay_reason">Reason: </label>
+                        <label for="open_cg_reason">Reason: </label>
                     </div>
                     <div class="col-md-4">
-                        <select name="stay_reason" v-model="formdata.reason" class="form-control" id="stay_reason">
-                            <option value="1">Reason 1</option>
-                            <option value="2">Reason 2</option>
-                            <option value="3">Reason 3</option>
-                            <option value="other">Other</option>
+                        <select name="closure_reason" v-model="formdata.reason" class="form-control" id="close_cg_reason">
+                            <option value="1">Closed due to natural disaster</option>
+                            <option value="2">Closed for maintenance</option>
+                            <option value="3">Other</option>
                         </select>
                     </div>
                 </div>
@@ -62,10 +51,10 @@
             <div v-show="requireDetails" class="row">
                 <div class="form-group">
                     <div class="col-md-2">
-                        <label for="stay_details">Details: </label>
+                        <label for="open_cg_details">Details: </label>
                     </div>
                     <div class="col-md-5">
-                        <textarea name="stay_details" v-model="formdata.details" class="form-control" id="stay_details"></textarea>
+                        <textarea name="closure_details" v-model="formdata.details" class="form-control" id="close_cg_details"></textarea>
                     </div>
                 </div>
             </div>
@@ -76,33 +65,27 @@
 </template>
 
 <script>
-import bootstrapModal from '../utils/bootstrap-modal.vue'
-import {bus} from '../utils/eventBus.js'
-import { $, datetimepicker,api_endpoints, validate, helpers } from '../../hooks'
-import alert from '../utils/alert.vue'
+import bootstrapModal from '../../utils/bootstrap-modal.vue'
+import {bus} from '../../utils/eventBus.js'
+import { $, datetimepicker,api_endpoints, validate, helpers } from '../../../hooks'
+import alert from '../../utils/alert.vue'
 module.exports = {
-    name: 'addMaxStayCS',
-    props: {
-        campsite: {
-            type: Object,
-            required: true
-        }
-    },
+    name: 'pkCsClose',
     data: function() {
         return {
+            status: '',
+            id:'',
             formdata: {
-                max_days: '',
                 range_start: '',
                 range_end: '',
                 reason:'',
                 details: ''
             },
-            start_picker: '',
-            end_picker: '',
+            closeStartPicker: '',
+            closeEndPicker: '',
             errors: false,
             errorString: '',
-            form: '',
-            isOpen: false
+            form: ''
         }
     },
     computed: {
@@ -111,10 +94,10 @@ module.exports = {
             return vm.errors;
         },
         isModalOpen: function() {
-            return this.isOpen;
+            return this.$parent.isOpenCloseCS;
         },
         requireDetails: function () {
-            return (this.formdata.reason === 'other')? true: false;
+            return (this.formdata.reason === '3')? true: false;
         }
     },
     components: {
@@ -123,10 +106,10 @@ module.exports = {
     },
     methods: {
         close: function() {
-            this.isOpen = false;
+            this.$parent.isOpenCloseCS = false;
             this.status = '';
         },
-        addMaxStay: function() {
+        addClosure: function() {
             if (this.form.valid()){
                 this.sendData();
             }
@@ -134,42 +117,41 @@ module.exports = {
         sendData: function() {
             let vm = this;
             var data = this.formdata;
-            data.campsite = vm.campsite.id;
+            data.status = vm.formdata.reason;
             $.ajax({
-                url: api_endpoints.campsites_stay_history,
+                url: api_endpoints.opencloseCS(vm.id),
                 method: 'POST',
                 xhrFields: { withCredentials:true },
                 data: data,
                 dataType: 'json',
                 success: function(data, stat, xhr) {
                     vm.close();
-                    vm.$parent.refreshMaxStayTable();
+                    bus.$emit('refreshCSTable');
                 },
                 error:function (resp){
                     vm.errors = true;
                     vm.errorString = helpers.apiError(resp);
                 }
             });
-
         },
         addFormValidations: function() {
             let vm = this;
             this.form.validate({
                 rules: {
-                    stay_start: "required",
-                    stay_reason: "required",
-                    stay_details: {
+                    closure_start: "required",
+                    closure_reason: "required",
+                    closure_details: {
                         required: {
                             depends: function(el){
-                                return vm.formdata.reason === 'other';
+                                return vm.formdata.reason === '3';
                             }
                         }
                     }
                 },
                 messages: {
-                    stay_start: "Enter a start date",
-                    stay_reason: "Select an open reason from the options",
-                    open_details: "Details required if Other reason is selected"
+                    closure_start: "Enter a start date",
+                    closure_reason: "Select a closure reason from the options",
+                    closure_details: "Details required if Other reason is selected"
                 },
                 showErrors: function(errorMap, errorList) {
 
@@ -197,21 +179,28 @@ module.exports = {
     },
     mounted: function() {
         var vm = this;
-        vm.start_picker = $('#stay_start_picker');
-        vm.end_picker = $('#stay_end_picker');
-        vm.start_picker.datetimepicker({
-            format: 'DD/MM/YYYY'
+        bus.$on('opencloseCS', function(data){
+            vm.status = data.status;
+            vm.id = data.id;
         });
-        vm.end_picker.datetimepicker({
-            format: 'DD/MM/YYYY'
+        vm.closeStartPicker = $('#close_cg_range_start');
+        vm.closeEndPicker = $('#close_cg_range_end');
+        vm.closeStartPicker.datetimepicker({
+            format: 'DD/MM/YYYY',
+            minDate: new Date()
         });
-        vm.start_picker.on('dp.change', function(e){
-            vm.formdata.range_start = vm.start_picker.data('DateTimePicker').date().format('DD/MM/YYYY');
+        vm.closeEndPicker.datetimepicker({
+            format: 'DD/MM/YYYY',
+            useCurrent: false
         });
-        vm.end_picker.on('dp.change', function(e){
-            vm.formdata.range_end = vm.end_picker.data('DateTimePicker').date().format('DD/MM/YYYY');
+        vm.closeStartPicker.on('dp.change', function(e){
+            vm.formdata.range_start = vm.closeStartPicker.data('DateTimePicker').date().format('DD/MM/YYYY');
+            vm.closeEndPicker.data("DateTimePicker").minDate(e.date);
         });
-        vm.form = $('#addMaxStayForm');
+        vm.closeEndPicker.on('dp.change', function(e){
+            vm.formdata.range_end = vm.closeEndPicker.data('DateTimePicker').date().format('DD/MM/YYYY');
+        });
+        vm.form = $('#closeCGForm');
         vm.addFormValidations();
     }
 };
