+++ parkstay/frontend/parkstay/src/components/campsites/closureHistory/openCampsite.vue
@@ -61,10 +61,10 @@
 </template>
 
 <script>
-import bootstrapModal from '../utils/bootstrap-modal.vue'
-import {bus} from '../utils/eventBus.js'
-import { $, datetimepicker,api_endpoints, validate, helpers } from '../../hooks'
-import alert from '../utils/alert.vue'
+import bootstrapModal from '../../utils/bootstrap-modal.vue'
+import {bus} from '../../utils/eventBus.js'
+import { $, datetimepicker,api_endpoints, validate, helpers } from '../../../hooks'
+import alert from '../../utils/alert.vue'
 module.exports = {
     name: 'pkCsOpen',
     data: function() {
